reorganize upload code in prep for merge logic, add more tests
