+++ src/upload/index.ts
@@ -1,0 +1,6 @@
+import * as core from '@actions/core'
+import {run} from './upload-artifact'
+
+run().catch(error => {
+  core.setFailed((error as Error).message)
+})
